// src/main/ts/core/types.ts
export enum ToggleMethods {
  on = "on",
  ON = "ON",
  off = "off",
  OFF = "OFF",
  toggle = "toggle",
  TOGGLE = "TOGGLE",
  determinate = "determinate",
  DETERMINATE = "DETERMINATE",
  indeterminate = "indeterminate",
  INDETERMINATE = "INDETERMINATE",
  enable = "enable",
  ENABLE = "ENABLE",
  disable = "disable",
  DISABLE = "DISABLE",
  readonly = "readonly",
  READONLY = "READONLY",
  destroy = "destroy",
  DESTROY = "DESTROY",
  rerender = "rerender",
  RENDERER = "RENDERER"
}